main

+++ controllers/product.controller.js
@@ -1,5 +1,3 @@
-const fs = require('fs')
-const db = require('../config/db');
 const Product = require("../models/index").Product;
 const User = require("../models/index").User;
 const jwt = require('jsonwebtoken');
@@ -21,33 +19,95 @@ exports.getProduct = async (req, res) => {
 }
 
 exports.postProduct = async (req, res) => {
-    const CategoryId = req.id;
     const body = req.body;
+    const CategoryId = body.CategoryId;
     const title = body.title;
     const price = body.price;
     const stock = body.stock;
+    const userRole = req.role
+    if(userRole == "admin"){
+        await Product.create({
+            title: title,
+            price: price,
+            stock: stock,
+            CategoryId : CategoryId
+        })
+        .then(product => {
+            res.status(201).json({
+               id: product.id,
+               title: product.title,
+               price: product.price,
+               stock: product.stock,
+               data : product
+            })
+        }).catch(e => {
+            console.log(e)
+            res.status(500).send({
+                status : "FAIL",
+                message : "Gagal membuat Product"
+            })
+        })
+    } else {
+        res.status(401).send({
+            message : "Only admin can access"
+        })
+    }
+    
+}
 
-    await Product.create({
-        title: title,
-        price: price,
-        stock: stock,
-        CategoryId : CategoryId
+exports.patchProduct = async (req, res) => {
+    const id = req.params.productId
+    const body = req.body
+    const userRole = req.role
+    if(userRole == "admin"){
+        await Product.findOne({
+        where: {
+            id: id
+        }
+    }).then( checkProduct => {
+        if (!checkProduct) {
+            res.status(404).send({
+                status: 404,
+                message: `Product by id '${id}' is not found`
+            })
+        } else {
+            Product.update({
+                CategoryId: body.CategoryId
+            }, {
+                where: {
+                    id: id
+                }
+            }).then( product => {
+                if (!product) {
+                    res.status(400).send({
+                        status: 400,
+                        message: 'Failed for update Product'
+                    })
+                } else {
+                    Product.findOne({
+                        where : {
+                            id : id
+                        }
+                    }).then(product => {
+                        res.status(200).send({
+                            "Product": product
+                        })
+                    })      
+                }
+            }).catch( err => {
+                console.log(err)
+                res.status(503).send({
+                    status: 503,
+                    message: 'Internal server error'
+                })
+            })
+        }
     })
-    .then(product => {
-        res.status(201).json({
-           id: product.id,
-           title: product.title,
-           price: product.price,
-           stock: product.stock,
-           data : product
+    }else {
+        res.status(401).send({
+            message : "Only admin can access"
         })
-    }).catch(e => {
-        console.log(e)
-        res.status(500).send({
-            status : "FAIL",
-            message : "Gagal membuat Product"
-        })
-    })
+    }
 }
 
 exports.updateProduct = async (req, res) => {
@@ -60,40 +120,54 @@ exports.updateProduct = async (req, res) => {
         price,
         stock,
     };
-    await Product.update(dataProduct, {
-        where : { id: productId},
-        returning: true,
-    })
-    .then((product) => {
-        res.status(200).json({
-            product: product[1]
+    const userRole = req.role
+    if(userRole == "admin"){
+        await Product.update(dataProduct, {
+            where : { id: productId},
+            returning: true,
         })
-    })
-    .catch((error) => {
-        console.log(error);
-        res.status(503).json({
-            message: "INTERNAL SERVER",
-            error :error,
-        });
-    })
+        .then((product) => {
+            res.status(200).json({
+                product: product[1]
+            })
+        })
+        .catch((error) => {
+            console.log(error);
+            res.status(503).json({
+                message: "INTERNAL SERVER",
+                error :error,
+            });
+        })
+    }else {
+        res.status(401).send({
+            message : "Only admin can access"
+        })
+    }
 }
-
 
 exports.deleteProduct = (req, res) => {
     const productId = req.params.productId;
-    Product.destroy({
-      where: {  id: productId },
-    })
-    .then (() => {
-        res.status(200).json({
-            message: "Your Photo has been succesfully deleted",
-        });
-    })
-    .catch((error) => {
-        console.log(error);
-        res.status(500).json({
-            message: "INTERNAL SERVER",
-            error: error,
-        });
-    })
+    const userRole = req.role
+    if(userRole == "admin"){
+        Product.destroy({
+            where: {  id: productId, Ca },
+          })
+          .then (() => {
+              res.status(200).json({
+                  message: "Your Photo has been succesfully deleted",
+              });
+          })
+          .catch((error) => {
+              console.log(error);
+              res.status(500).json({
+                  message: "INTERNAL SERVER",
+                  error: error,
+              });
+          })
+    }else {
+        res.status(401).send({
+            message : "Only admin can access"
+        })
+    }
+    
 }
